repository

+++ fulfillment/src/main/java/com/springboot/fulfillment/data/repository/DeliveryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springboot.fulfillment.data.entity.Delivery;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
 }
